do not remove comments blindly

+++ codeInsight/impl/com/intellij/codeInsight/daemon/impl/RemoveSuppressWarningAction.java
@@ -9,9 +9,9 @@ import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.psi.*;
-import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.psi.javadoc.PsiDocComment;
 import com.intellij.psi.javadoc.PsiDocTag;
+import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.ArrayUtil;
 import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.NotNull;
@@ -73,10 +73,10 @@ public class RemoveSuppressWarningAction implements LocalQuickFix {
 
   private void removeFromComment(final PsiComment comment) throws IncorrectOperationException {
     String newText = removeFromElementText(comment);
-    if (newText == null) {
+    if (newText != null && newText.length() == 0) {
       comment.delete();
     }
-    else {
+    else if (newText != null) {
       PsiComment newComment = comment.getManager().getElementFactory().createCommentFromText("// " + GlobalInspectionContextImpl.SUPPRESS_INSPECTIONS_TAG_NAME+" "+newText, comment);
       comment.replace(newComment);
     }
@@ -86,10 +86,10 @@ public class RemoveSuppressWarningAction implements LocalQuickFix {
     PsiDocTag tag = docComment.findTagByName(GlobalInspectionContextImpl.SUPPRESS_INSPECTIONS_TAG_NAME);
     if (tag == null) return;
     String newText = removeFromElementText(tag.getDataElements());
-    if (newText == null) {
+    if (newText != null && newText.length() == 0) {
       tag.delete();
     }
-    else {
+    else if (newText != null) {
       newText = "@" + GlobalInspectionContextImpl.SUPPRESS_INSPECTIONS_TAG_NAME + " " + newText;
       PsiDocTag newTag = tag.getManager().getElementFactory().createDocTagFromText(newText, tag);
       tag.replace(newTag);
@@ -106,7 +106,7 @@ public class RemoveSuppressWarningAction implements LocalQuickFix {
     text = StringUtil.trimStart(text, GlobalInspectionContextImpl.SUPPRESS_INSPECTIONS_TAG_NAME).trim();
     List<String> ids = StringUtil.split(text, ",");
     int i = ArrayUtil.find(ids.toArray(), myID);
-    if (i==-1 || ids.size()==1) return null;
+    if (i==-1) return null;
     ids.remove(i);
     return StringUtil.join(ids, ",");
   }
